requirements.txt
gradio
python-dotenv
langchain-community
langchain
langchain-groq
langchain_google_genai
faiss-cpu
numpy
pypdf